Moving appbuilder.xxx out of view files ands into app.py
