+++ superset/views/database/api.py
@@ -56,6 +56,3 @@ class DatabaseRestApi(DatabaseMixin, ModelRestApi):
     # Removes the local limit for the page size
     max_page_size = -1
     validators_columns = {"sqlalchemy_uri": sqlalchemy_uri_validator}
-
-
-appbuilder.add_api(DatabaseRestApi)
