+++ superset/views/sql_lab.py
@@ -21,10 +21,10 @@ from flask import g, redirect, request, Response
 from flask_appbuilder import expose
 from flask_appbuilder.models.sqla.interface import SQLAInterface
 from flask_appbuilder.security.decorators import has_access, has_access_api
-from flask_babel import gettext as __, lazy_gettext as _
+from flask_babel import lazy_gettext as _
 from flask_sqlalchemy import BaseQuery
 
-from superset import appbuilder, db, get_feature_flags, security_manager
+from superset import db, get_feature_flags, security_manager
 from superset.models.sql_lab import Query, SavedQuery, TableSchema, TabState
 from superset.utils import core as utils
 
@@ -69,16 +69,6 @@ class QueryView(SupersetModelView):
         "start_time": _("Start Time"),
         "end_time": _("End Time"),
     }
-
-
-appbuilder.add_view(
-    QueryView,
-    "Queries",
-    label=__("Queries"),
-    category="Manage",
-    category_label=__("Manage"),
-    icon="fa-search",
-)
 
 
 class SavedQueryView(
@@ -171,10 +161,6 @@ class SavedQueryViewApi(SavedQueryView):  # pylint: disable=too-many-ancestors
     @expose("show/<pk>")
     def show(self, pk):
         return super().show(pk)
-
-
-appbuilder.add_view_no_menu(SavedQueryViewApi)
-appbuilder.add_view_no_menu(SavedQueryView)
 
 
 def _get_owner_id(tab_state_id):
@@ -332,15 +318,6 @@ class TableSchemaView(BaseSupersetView):
         return json_success(response)
 
 
-appbuilder.add_view_no_menu(TabStateView)
-appbuilder.add_view_no_menu(TableSchemaView)
-
-
-appbuilder.add_link(
-    __("Saved Queries"), href="/sqllab/my_queries/", icon="fa-save", category="SQL Lab"
-)
-
-
 class SqlLab(BaseSupersetView):
     """The base views for Superset!"""
 
@@ -349,6 +326,3 @@ class SqlLab(BaseSupersetView):
     def my_queries(self):  # pylint: disable=no-self-use
         """Assigns a list of found users to the given role."""
         return redirect("/savedqueryview/list/?_flt_0_user={}".format(g.user.id))
-
-
-appbuilder.add_view_no_menu(SqlLab)
